backend/requirements.txt
deepface
Pillow
numpy
